app, pdb: silence coverity warning about integer overflow
Brushes can't be large enough to actually cause overflow here, but
since the result is gsize, it can't hurt to have the intermediate
results to gsize too.

+++ app/pdb/brush-cmds.c
@@ -230,7 +230,7 @@ brush_get_pixels_invoker (GimpProcedure         *procedure,
       width          = gimp_temp_buf_get_width  (mask);
       height         = gimp_temp_buf_get_height (mask);
       mask_bpp       = babl_format_get_bytes_per_pixel (format);
-      num_mask_bytes = gimp_temp_buf_get_height (mask) *
+      num_mask_bytes = (gsize) gimp_temp_buf_get_height (mask) *
                        gimp_temp_buf_get_width  (mask) * mask_bpp;
       mask_bytes     = g_bytes_new (data, num_mask_bytes);
 
@@ -243,7 +243,7 @@ brush_get_pixels_invoker (GimpProcedure         *procedure,
           data   = gimp_temp_buf_lock (pixmap, format, GEGL_ACCESS_READ);
 
           color_bpp       = babl_format_get_bytes_per_pixel (format);
-          num_color_bytes = gimp_temp_buf_get_height (pixmap) *
+          num_color_bytes = (gsize) gimp_temp_buf_get_height (pixmap) *
                             gimp_temp_buf_get_width  (pixmap) *
                             color_bpp;
           color_bytes     = g_bytes_new (data, num_color_bytes);
